requirements.txt
absl-py==1.3.0
altair==4.2.0
aniso8601==9.0.1
anyascii==0.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astor==0.8.1
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitstring==4.0.2
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
ConfigArgParse==1.5.3
contourpy==1.0.6
contractions==0.1.73
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
ecdsa==0.18.0
entrypoints==0.4
esptool==4.5.1
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
filelock==3.8.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flatbuffers==22.11.23
fonttools==4.38.0
fuzzywuzzy==0.18.0
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.9
gitdb==4.0.10
GitPython==3.1.29
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
httplib2==0.21.0
huggingface-hub==0.13.4
idna==3.3
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.1.0
imutils==0.5.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter-http-over-ws==0.0.8
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.20.9
libclang==14.0.6
librosa==0.10.0.post2
llvmlite==0.39.1
locust==2.15.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
msgpack==1.0.5
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pafy==0.5.5
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.0
pexpect==4.8.0
phonenumbers==8.13.1
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2022.11.25
platformdirs==2.5.4
pooch==1.6.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==1.4.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.2.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
rapidfuzz==2.15.0
redis==4.5.4
reedsolo==1.6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
roundrobin==0.0.4
rsa==4.9
scikeras==0.9.0
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
selenium==4.7.2
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.98
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.4
spotipy==2.22.1
SQLAlchemy==2.0.8
sqlparse==0.4.3
stack-data==0.6.1
stevedore==4.1.1
streamlit==1.15.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
termcolor==2.1.1
terminado==0.17.0
textsearch==0.0.24
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
validators==0.20.0
virtualenv==20.17.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
youtube-dl==2021.12.17
zipp==3.11.0
zope.event==4.6
zope.interface==6.0
